refactor: chunkの処理を分割
